requirements.txt
Django==3.0.5
gunicorn
django-widget-tweaks==1.4.8
sqlparse==0.3.1
django-import-export
djangorestframework
markdown
django-filter
pandas
whitenoise
virtualenv==20.0.34
Pillow
